Reorganize EigenDA client codecs
