+++ api/clients/codecs/encodings_test.go
@@ -1,10 +1,12 @@
-package clients
+package codecs_test
 
 import (
 	"bytes"
 	"crypto/rand"
 	"math/big"
 	"testing"
+
+	"github.com/Layr-Labs/eigenda/api/clients/codecs"
 )
 
 // Helper function to generate a random byte slice of a given length
@@ -20,7 +22,7 @@ func randomByteSlice(length int64) []byte {
 // TestDefaultBlobEncodingCodec tests the encoding and decoding of random byte streams
 func TestDefaultBlobEncodingCodec(t *testing.T) {
 	// Create an instance of the DefaultBlobEncodingCodec
-	codec := DefaultBlobEncodingCodec{}
+	codec := codecs.DefaultBlobEncodingCodec{}
 
 	// Number of test iterations
 	const iterations = 100
@@ -53,7 +55,7 @@ func TestDefaultBlobEncodingCodec(t *testing.T) {
 }
 
 func TestIFFTBlobEncodingCodec(t *testing.T) {
-	codec := IFFTBlobEncodingCodec{}
+	codec := codecs.IFFTBlobEncodingCodec{}
 
 	// Number of test iterations
 	const iterations = 100
