+++ api/clients/eigenda_client_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/Layr-Labs/eigenda/api/clients"
+	"github.com/Layr-Labs/eigenda/api/clients/codecs"
 	clientsmock "github.com/Layr-Labs/eigenda/api/clients/mock"
 	"github.com/Layr-Labs/eigenda/api/grpc/common"
 	grpcdisperser "github.com/Layr-Labs/eigenda/api/grpc/disperser"
@@ -65,10 +66,10 @@ func TestPutRetrieveBlobSuccess(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	expectedBlob := []byte("dc49e7df326cfb2e7da5cf68f263e1898443ec2e862350606e7dfbda55ad10b5d61ed1d54baf6ae7a86279c1b4fa9c49a7de721dacb211264c1f5df31bade51c")
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), expectedBlob)
@@ -96,10 +97,10 @@ func TestPutBlobFailDispersal(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 	require.Error(t, err)
@@ -128,10 +129,10 @@ func TestPutBlobFailureInsufficentSignatures(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 	require.Error(t, err)
@@ -160,10 +161,10 @@ func TestPutBlobFailureGeneral(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 	require.Error(t, err)
@@ -192,10 +193,10 @@ func TestPutBlobFailureUnknown(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 	require.Error(t, err)
@@ -226,10 +227,10 @@ func TestPutBlobFinalizationTimeout(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 	require.Error(t, err)
@@ -285,10 +286,10 @@ func TestPutBlobIndividualRequestTimeout(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
 
@@ -347,10 +348,10 @@ func TestPutBlobTotalTimeout(t *testing.T) {
 			CustomQuorumIDs:          []uint{},
 			SignerPrivateKeyHex:      "75f9e29cac7f5774d106adb355ef294987ce39b7863b75bb3f2ea42ca160926d",
 			DisableTLS:               false,
-			PutBlobEncodingVersion:   clients.DefaultBlobEncoding,
+			PutBlobEncodingVersion:   codecs.DefaultBlobEncoding,
 		},
 		Client:   disperserClient,
-		PutCodec: clients.DefaultBlobEncodingCodec{},
+		PutCodec: codecs.DefaultBlobEncodingCodec{},
 	}
 	blobInfo, err := eigendaClient.PutBlob(context.Background(), []byte("hello"))
